Redesign and continue build
In the Chromosome class, I redesign the mutation method and simply it. In the main VRP_GA class, I continue the algorithm, now, 1 vehicle and less than 50 locations mode could be used for shortest path. Next step for this algorithm is to use the simulate data to test the first version of the mode1.