delete useless code and resolve errors
In the main VRP_GA.java file, delete some useless code in the begining code stage. after that, all the red error in the Chromosome methods and VRP_GA class mode1. All of these method could use in the all modes. Next step for the code is to solve the logical problem in the loop.